[FRONTEND] `triton.jit.version_key()` more structured implementation + shorter hash (#1418)
May I suggest this? I wrote it on 'main' but it seems useful for a
refactor involving multiple files.

It's a little less hardcoded, so more easily adaptable for future
extension.
Also, it folds the hashing of all files into a single digest, rather
than concat-ing one hash per file, which creates unwieldy cache keys,
esp when subpackages are involved. These are visible at the very least
in every error traceback.
I wasn't sure whether the intention was to leave a visible hash for each
contained file. I assumed it wasn't.

More generally, I'm not sure whether the approach of hashing source code
and binaries is maintainable, portable or user-friendly in the long run.
It certainly doesn't help with the runtime of `import triton` (compare
to `import torch`, or `import numpy` which are near immediate for much
larger frameworks). Long import / initialization times encumber small
utility scripts / subprocesses that are frequently invoked in automated
workflows.